urbannest/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
idna==3.10
pillow==11.3.0
psycopg2==2.9.10
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
